Caracters finalized

+++ src/Controller/BestiaryController.php
@@ -11,6 +11,7 @@ use App\Form\RegistrationFormType;
 
 use App\Repository\WikiRepository;
 use App\Security\Voter\WikiElementVoter;
+use App\Security\Voter\WikiVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Serializer\SerializerInterface;
@@ -24,8 +25,8 @@ class BestiaryController extends AbstractController
 {
 
     #[Route('/api/wikis/{wiki}/bestiaries', name: 'add.bestiary', methods: ['POST'])]
-    #[IsGranted(WikiElementVoter::EDIT, subject: 'bestiary')]
-    public function addbestiary(Bestiary $bestiary, Wiki $wiki, Request $request, BestiaryRepository $bestiaryRepository, WikiRepository $wikiRepository,SerializerInterface $serializer): Response
+    #[IsGranted(WikiVoter::EDIT, subject: 'wiki')]
+    public function addbestiary(Wiki $wiki, Request $request, BestiaryRepository $bestiaryRepository, WikiRepository $wikiRepository,SerializerInterface $serializer): Response
     {
         $bestiary = $serializer->deserialize($request->getContent(), bestiary::class, 'json');
         $bestiary->setWiki($wiki);
@@ -41,6 +42,11 @@ class BestiaryController extends AbstractController
     #[IsGranted(WikiElementVoter::EDIT, subject: 'bestiary')]
     public function deletebestiary(Bestiary $bestiary, Wiki $wiki, bestiaryRepository $bestiaryRepository, WikiRepository $wikiRepository): Response
     {
+        // Vérifiez si le bestiary appartient bien au wiki
+        if ($bestiary->getWiki()->getId() !== $wiki->getId()) {
+            return $this->json(['error' => 'Bestiary does not belong to the specified wiki.'], 400);
+        }
+
         $bestiaryRepository->removeBestiary($bestiary, $wiki);
         $wiki = $wikiRepository->findOneById($wiki->getId());
         return $this->json($wiki, 200, [], [
@@ -49,9 +55,26 @@ class BestiaryController extends AbstractController
     }
 
     #[Route('/api/wikis/{wiki}/bestiaries/{bestiary}', name: 'update.bestiary', methods: ['PUT'])]
-    #[IsGranted(WikiElementVoter::EDIT, subject: 'bestiary')]
-    public function updateBestiary(Bestiary $bestiary, Wiki $wiki, Request $request, BestiaryRepository $bestiaryRepository, WikiRepository $wikiRepository, SerializerInterface $serializer): Response
+    #[IsGranted(WikiVoter::EDIT, subject: 'wiki')]
+    public function updateBestiary(Wiki $wiki, Bestiary $bestiary, Request $request, BestiaryRepository $bestiaryRepository, WikiRepository $wikiRepository, SerializerInterface $serializer): Response
     {
+        // Récupérer l'entité Wiki
+        $wiki = $wikiRepository->find($wiki);
+        if (!$wiki) {
+            return $this->json(['error' => 'Wiki not found.'], 404);
+        }
+    
+        // Récupérer l'entité Job
+        $job = $bestiaryRepository->find($bestiary);
+        if (!$job) {
+            return $this->json(['error' => 'Bestiary not found.'], 404);
+        }
+    
+        // Vérifiez si le job appartient bien au wiki
+        if ($job->getWiki()->getId() !== $wiki->getId()) {
+            return $this->json(['error' => 'Bestiary does not belong to the specified wiki.'], 400);
+        }
+
         $bestiary = $serializer->deserialize($request->getContent(), Bestiary::class, 'json', ['object_to_populate' => $bestiary]);
         $bestiaryRepository->updateBestiary($bestiary, $wiki);
